Create shared AppWidget application
AppWidget applications need to be stored internally or their
AppWidgets fail to load. So this application needs to be installed
separately and internally, which means it must be small.

On the other hand, it's not MLP-specific, so should be able to
create widgets for any soundboard which implements the same Intents
and ContentProvider columns.

It's obviously very rough, but is basically correct in framework
terms. Assets for the AppWidget should be supplied by the
soundboard via the ContentProvider interface later, so this
package can remain tiny and generic.