Librarise the Provider Columns and media player
The media player service will happily play any content URL its given.

I could have push the chooser Activity into the library, but it's the
part most likely to be customised per-soundboard.

The ContentProvider is realy the heart of the soundboard, so it cannot
be librarised. Not every soundboard is going to be playing things from
an asset collection in a two-level file namespace... Also, it needs to
know the MIME types and such which are application-specifc.

Neither of these leap out to me as needing a library superclass.