Added some category icons by Equestria-Election
Character icons by Equestria-Election (http://equestria-election.deviantart.com/)
Used with permission of artist.

It's only partial coverage, but serves to test and demonstrate
the icon support. And they're really cool!

I resized them down to 144 on the longest dimension using GIMP,
which is correct for XHDPI devices (72dp is the guaranteed minimum
size of an AppWidget slot, and XHDPI is 2:1 ). It looks like Android resizes
them OK for my MHDPI device, but a proper build pipeline would vastly
improve the process.