Crop all the non-headshots down during processing
I haven't checked specifically that all the relevant artists are
alright with this, but I don't think it's against the spirit of my
original request either. None of the artists have strongly qualified
my use of these assets (apart from a couple of "don't charge" and
the obvious Creative Commons license limitations) so I don't expect
this to be a problem.