Use XML to create IDs rather than asset-scans
This allows the IDs for specific sounds to remain stable as new audio
is added and old audio removed.

It has the added benefit of (on my HTC Dream, anyway) appearing to
start a bit faster.

However, the query list isn't sorted anymore, so the sound list is
pretty much in random at the momemnt.