Ensure ICON content:// uri ids are fixed
During the implementation of the category support in the SoundProvider,
I changed the ICON content:// uri ids to be the resource id of the
relevant drawable.

However, those numbers change when new drawables are added. And the
AppWidget gave that content:// uri to its RemoteImageView, which is
apparently seeing a "content provider updated" message sent when
a new version of the package is installed, and refreshing its icon
with the old content:// uri.

This produces an incorrect icon when new drawables are added.

So go back to the fixed ids, except we can accept a category id
(<1000) rather than a sound id, and hence still avoid any aliasing
in the content:// uris.