Fix playback cutting off too early
On examination, there were two problems here:

Under memory pressure, GC was being run, and noticing that we did
not have a reference to our MediaPlayer objects, was destroying them.
This could be triggered by scrolling around while playing.
Using prepareAsync seemed to be solving this on my HTC Dream perhaps
because the way it was streaming the audio in during the preparation
into presumably smaller buffers than more recent devices meant the
media player was still referenced.

The way I'd set up the OnCompletionLister meant that the only
MediaPlayer reference being held was in an object itself only
referenced by the MediaPlayer.

There's also a chance that I was calling stopSelfResult too early if
a short sound was started after a long sound, and newer devices were
perhaps garbage collecting sooner and killing the service and its
player... But I'm not sure. It's a bug, but I haven't been able to
construct a scenario where it'd kill the sound on newer devices than
my HTC Dream.