Delay MediaPlayer.release by 500ms
From user reports and experimentally with a Bluetooth headset,
it seems a MediaPlayer's OnCompletionListener can be called before
the sound is actually complete (probably once the hardware has
acknowledged the last packet), but MediaPlayer.release() is able to
immediately kill the playback in progress.

This was causing the sound to be audibly cut off before completion.

This appears to be a driver-dependant issue, as the same device that
shows the problem on a Bluetooth headset shows no such problem using
its internal speakers.

I took the opportunity to refactor the SoundboardService to not need
a Handler subclass, but simply to directly queue up two types of
runnables.