Synchronise all code that writes to the databases
This should fix the problem of potentially having two threads
trying to populate the database simultaneously, which will cause the
population code to fail.

Because data is present when this happens, the caller gets a "success"
message back, even though it actually failed badly.

This was potentially being triggered by the "content pack installed"
notification code, if a second pack finished installing and raised
the notification while the first was still scanning.