Change database format to version 3
This isn't actually a change in database format, but before
commit release-014-3-g06c7c8c it was possible to have mispopulated
the database due to mutiple threads running loadData().

If a Content Pack provides bad data due to its own missing locks,
the bad data will be refetched again, but that should be cleared
away in the next update to that Content Pack.